dfir_pipes::pull

Trait FusedPull

pub trait FusedPull: Pull { }

A marker trait for pulls that are “fused”.

A fused pull guarantees that once it returns PullStep::Ended, all subsequent calls to Pull::pull will also return PullStep::Ended. This property allows downstream operators like Pull::chain to avoid tracking whether the upstream has ended.

Implementors should ensure this invariant is upheld. The Pull::fuse adapter can be used to make any pull fused.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.