dfir_pipes

Type Alias No
·

pub type No = !;

Type-level false for Toggle.

Indicates that a capability is absent (e.g., the pull cannot pend or cannot end).

A type alias for core::convert::Infallible, representing a type that can never be constructed.

Used in Step variants that are statically impossible (e.g., Pending when CanPend = No). The error type for errors that can never happen.

Since this is an alias to the never type, a value of this type can never actually exist. This can be useful for generic APIs that use Result and parameterize the error type, to indicate that the result is always Ok.

For example, the TryFrom trait (conversion that returns a Result) has a blanket implementation for all types where a reverse Into implementation exists.

impl<T, U> TryFrom<U> for T where U: Into<T> {
    type Error = Infallible;

    fn try_from(value: U) -> Result<Self, Infallible> {
        Ok(U::into(value))  // Never returns `Err`
    }
}

Note: since CURRENT_RUSTC_VERSION this is an alias to !. If targeting that or future versions, prefer using the never type directly.